scripts/requirements.txt
sympy
open3d
numpy
pandas
mujoco
scikit-robot
urdfpy
trimesh
scipy
rospkg
matplotlib
pyyaml
scikit-robot